CDR-2305 merge TemplateCacheService into TemplateServiceImp
