+++ service/src/main/java/org/ehrbase/repository/TemplateStoreRepository.java
@@ -31,7 +31,7 @@ import org.ehrbase.jooq.pg.tables.CompVersion;
 import org.ehrbase.jooq.pg.tables.CompVersionHistory;
 import org.ehrbase.jooq.pg.tables.TemplateStore;
 import org.ehrbase.jooq.pg.tables.records.TemplateStoreRecord;
-import org.ehrbase.service.TemplateCacheService;
+import org.ehrbase.service.TemplateServiceImp.TemplateMetaData;
 import org.ehrbase.service.TimeProvider;
 import org.ehrbase.util.UuidGenerator;
 import org.jooq.DSLContext;
@@ -53,7 +53,7 @@ public class TemplateStoreRepository {
         this.timeProvider = timeProvider;
     }
 
-    public TemplateCacheService.TemplateMetaData store(TemplateCacheService.TemplateMetaData templateData) {
+    public TemplateMetaData store(TemplateMetaData templateData) {
         TemplateStoreRecord templateStoreRecord = context.newRecord(TEMPLATE_STORE);
         templateStoreRecord.setId(UuidGenerator.randomUUID());
         setTemplateFields(templateData, templateStoreRecord, timeProvider);
@@ -61,7 +61,7 @@ public class TemplateStoreRepository {
         return buildMetadata(templateStoreRecord);
     }
 
-    public TemplateCacheService.TemplateMetaData update(TemplateCacheService.TemplateMetaData templateData) {
+    public TemplateMetaData update(TemplateMetaData templateData) {
         String templateId = templateData.meta().templateId();
         TemplateStoreRecord templateStoreRecord = context.selectFrom(TEMPLATE_STORE)
                 .where(TEMPLATE_STORE.TEMPLATE_ID.eq(templateId))
@@ -108,8 +108,8 @@ public class TemplateStoreRepository {
                         r.component1(), r.component2(), r.component3(), r.component4(), r.component5()));
     }
 
-    private static TemplateCacheService.TemplateMetaData buildMetadata(TemplateStoreRecord rec) {
-        return new TemplateCacheService.TemplateMetaData(
+    private static TemplateMetaData buildMetadata(TemplateStoreRecord rec) {
+        return new TemplateMetaData(
                 rec.getContent(),
                 new TemplateService.TemplateDetails(
                         rec.getId(),
@@ -152,7 +152,7 @@ public class TemplateStoreRepository {
      * @param templateIds
      * @return the templates
      */
-    public List<TemplateCacheService.TemplateMetaData> findByTemplateIds(String... templateIds) {
+    public List<TemplateMetaData> findByTemplateIds(String... templateIds) {
 
         if (templateIds.length == 0) return List.of();
 
@@ -176,9 +176,7 @@ public class TemplateStoreRepository {
     }
 
     private static void setTemplateFields(
-            TemplateCacheService.TemplateMetaData templateData,
-            TemplateStoreRecord templateStoreRecord,
-            TimeProvider timeProvider) {
+            TemplateMetaData templateData, TemplateStoreRecord templateStoreRecord, TimeProvider timeProvider) {
         templateStoreRecord.setTemplateId(templateData.meta().templateId());
         templateStoreRecord.setCreationTime(timeProvider.getNow());
         templateStoreRecord.setContent(templateData.operationalTemplate());
