+++ service/src/test/java/org/ehrbase/test/fixtures/TemplateFixture.java
@@ -25,17 +25,14 @@ import java.util.UUID;
 import org.apache.xmlbeans.XmlException;
 import org.ehrbase.api.service.TemplateService;
 import org.ehrbase.openehr.sdk.test_data.operationaltemplate.OperationalTemplateTestData;
-import org.ehrbase.service.TemplateCacheService;
+import org.ehrbase.service.TemplateServiceImp.TemplateMetaData;
 import org.ehrbase.util.TemplateUtils;
 import org.openehr.schemas.v1.OPERATIONALTEMPLATE;
 import org.testcontainers.shaded.org.apache.commons.io.IOUtils;
 
 public class TemplateFixture {
 
-    public record TestTemplate(
-            String templateId,
-            OPERATIONALTEMPLATE operationaltemplate,
-            TemplateCacheService.TemplateMetaData metaData) {}
+    public record TestTemplate(String templateId, OPERATIONALTEMPLATE operationaltemplate, TemplateMetaData metaData) {}
 
     public static TestTemplate fixtureTemplate(OperationalTemplateTestData operationalTemplateTestData) {
         return fixtureTemplate(operationalTemplateTestData, UUID.fromString("b65165e8-b4a6-4c23-84a0-ec58cfb481c1"));
@@ -57,7 +54,7 @@ public class TemplateFixture {
             throw new RuntimeException(e);
         }
 
-        TemplateCacheService.TemplateMetaData metaData = new TemplateCacheService.TemplateMetaData(
+        TemplateMetaData metaData = new TemplateMetaData(
                 templateDocument,
                 new TemplateService.TemplateDetails(
                         internalUUID,
